fix: move scrypt to utils
