+++ src/users/users.service.ts
@@ -3,21 +3,7 @@ import { CreateUserDTO } from './dto/create-user.dto';
 import { UpdatePasswordDTO } from './dto/update-password.dto';
 import { PrismaService } from '../database/prisma.service';
 import { errors } from '../common/utils/errors';
-import { scrypt, timingSafeEqual } from 'crypto';
-import { CRYPT_COST, CRYPT_SALT, KEY_LEN } from '../common/constants';
-
-export async function scryptHash(password: string) {
-  return new Promise<string>((resolve) =>
-    scrypt(password, CRYPT_SALT, KEY_LEN, { cost: 2 ** CRYPT_COST }, (_, key) =>
-      resolve(key.toString('base64')),
-    ),
-  );
-}
-
-export async function scryptCompare(password: string, hashed: string) {
-  const hashedPassword = await scryptHash(password);
-  return timingSafeEqual(Buffer.from(hashedPassword), Buffer.from(hashed));
-}
+import { scryptCompare, scryptHash } from '../common/utils/scrypt';
 
 @Injectable()
 export class UsersService {
